[FE][Fix][Refactor] : Layout 폴더 src로 이동 및 console.log 삭제
