+++ frontend/src/layout/footer/LayoutFooter.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
-import { FooterContext } from '@/component/layout/footer/LayoutFooterProvider';
-import { Footer } from '@/component/layout/footer/Footer';
+import { FooterContext } from '@/layout/footer/LayoutFooterProvider.tsx';
+import { Footer } from '@/layout/footer/Footer.tsx';
 
 export const LayoutFooter = () => {
   const { footerOption } = useContext(FooterContext);
